Initial commit
I'll admit, I probably should have committed sooner, but this is an
adaptation (read: complete rewrite) of what the project looked like
beforehand. Before, though, it was an absolute mess. Everything was
hacked together, and sure, it worked, but adding new features was a pain
in the ass.

I was motivated to do this rewrite through Andreas Kling's LibJS videos,
so this project follows a structure similar to SerenityOS's LibJS. In
this light, this project does not compile to native bytecode, but acts
more like an interpreter.

Part of the reason the project was so messy before this rewrite is
because we were trying to make it behave exactly like Rhino (of which we are
very familiar with). The problem, though, is that Rhino is such an old
project and isn't even written in the same language. It is messy, uses
buggy first-party implementations of things that could easy be done from
a library, and has many inconsistencies. It works well, and I had fun
messing around with it, but it should not serve as the base for a new
project.

Now, on to the details of this commit.

First of all, unlike Rhino, we make use of Antlr4 to do our lexing and
parsing. We use a modified version of the .g4 file provided in the
antlr/grammars-v4 repository. In addition to renaming certain rules to
make more sense, the only modifications that have been made to this file
have been to split up rules into different names to make handling it in the
Kotlin side easier. For example, take the following rule:

    bitShiftExpr: arithmeticExpr (('<<' | '>>') arithmeticExpr)*;

While parsing the context produced by this rule, it is non-intuitive to
get the particular shift string used in this context. One is required to
get a token at a certain position, and it may not necessary be obvious
what the code is doing. Compare that to the following rules:

    bitShiftExpr: arithmeticExpr (bitShiftOperator arithmeticExpr)*;
    bitShiftOperator: '<<' | '>>';

When dealing with this code, one would call ctx.bitShiftOperator() to
get the TerminalNode, which is very clear as to the purpose of the code.

After lexing and parsing with Antlr, the context (currently only a
FileInputContext) is transformed via the IRVisitor to a tree of Nodes.
These nodes serve as the intermediate representation of the Python code.
The IRVisitor always produces a ScriptNode as the top level node in the
tree.

After transformation, the resulting ScriptNode is then passed to the
Interpreter, where each node is executed. Each node is responsible for
it's own execution. It may choose to simply delegate the execution
directly to it's children (ex: StatementNode), or it may choose to do
more complex logic (ex: FunctionNode). Delegating control over this
process to the individual Nodes breaks up the code into easy-to-manage
chunks. Additionally, this system opens up the door for the possibility
of JIT compilation later down the road. As each node is responsible for
it's own interpretation, it may also be responsible for it's own
compilation.