Deleted the TYPE convention
I had a fundamental misunderstanding about Python. I was executing
something like "type(object)", and when that printed "<class'type'>", I
interpretted that as saying that the object global was an instance of
the type class. In reality, "type" is a class, and that statement was
invoking the type constructor with "object" as it's argument. It
converted it to a type, hence the print output.

Damn you Python, and your lack of a "new" operator! :(

Now, though, the inheritance system is a lot nicer. I initialize methods
in the init blocks of the classes, and I don't need a separate inherit
method -- standard class inheritance does the job for me.