Complete refactor of the type system
Previously, I was struggling with a circularity issue of the python type
system. The fact that every object has a type, but the type is also an
object, threw me through a bit of a loop. How can I instantiate an
object that requires itself to create?

The solution is the make the type a hidden internal object that doesn't
inherit from VObject, which is being called a ClassDescriptor. This way,
every VObject can take a ClassDescriptor, which contains all of the
methods and properties, without running into circularity.

Additionally, this makes it really easy to extend from other classes.
Since a ClassDescriptor is really just a map, one can easily merge one
ClassDescriptor into another.

One thing to look into is that, when a property is added, if it is found
on the descriptor, it is changed on the descriptor. That probably
shouldn't happen, or at least not always. There may be certain times
where that is appropriate, but that will have to wait until the type
system is flushed out a bit more, and more magic methods are added with
correct implementations (like "__getattr__" & "__getattribute__").