VObject and ClassDescriptor now inherit from ParentDeferredMap
Previously, there was a lot of code duplication to get VObject and
ClassDescriptor to delegate to their parents if applicable. Here, a
simple ParentDeferredMap data structure was created which is a wrapper
around two maps, it's own map and a nullable parent map. It simply makes
all calls to it's map, and/or parent, if applicable.

This cleans up a lot of code regarding selecting descriptors for various
classes.