Improved flow statements
Flow statements are now handled very generally. With this commit comes a
new VObject, called VFlowWrapper, which can be returned from
Node#execute. If a node returns this, it is an indication to some parent
node to take some special action.

For example, if a FunctionNode executes it's children and one of them
returns a VReturnNode, we know that we've returned a value. Otherwise,
the FunctionNode will return VNone, as there was no return statement.

To facilitate this change, all nodes which simply execute a block of
statements now do so with StatementNode.executeStatements. This checks
each statement execution to see if it returned a VFlowWrapper, and if
so, the value is returned immediately and the rest of the statements are
not executed.