Creating VObjects create bound methods from the ClassDescriptor
Previously, the passing of self to an instance's method was hacked
together, simply being appended to the arguments list at the call site.
Now, in the VObject constructor, the object iterates through every
method available in it's descriptor, and create a bound function wrapped
in a LazyValue (so it doesn't actually create another VObject it the
constructor).

This does not take inheritance into account. In other words, if class A
defines method foo, and class B inherit from A, then when B is
instantiated, the method will be copied from A to B. This should not
happen, and method lookup should just delegate to A. Unfortunately, its
a bit hard to implement and test this before fleshing out the VType
class a bit more and adding the built-in type global ('object', 'int',
etc...).