Rewrote type system (hopefully for the last time)
Now that I have a better understanding of Python's type system, it was
time to rewrite Mamba's type system in order to support some of the
finer aspects, such as (seemingly) infinite property access (think
".__call__.__call__.__call__....").

ClassDescriptor has been replaced in favor of what I originally wanted
it to be: VType. Every object's type is now a VType (which is also an
object with it's own VType), which is how Python actually works. In
order to do this, VTypes are usually passed around wrapped in a
LazyValue<VType>, which is simply a producer for a VType.

This allows VType's type to be LazyValue { VType }, as long as no
VObject constructor ever attempts to unwrap a LazyValue.

Object methods are defined on the object's type, and bound (!!) to the
actual object on creation. Note that the magic methods for instantiation
have not been implemented.