Fix method prioritization during binding
Now, if the class has a method already present in it's map, and it finds
a method in one of it's base classes with the same name, the value in
the base class is ignored. Additionally, base classes are handled bottom
up now, meaning that a type's direct base get it's methods bound, then
that base gets it's base's methods bound, and so on.