Drastically simplied type system
Type methods are incomplete, and some things still need quite a bit of
work (like scope resolution), however it seems to function at a base
level now. VObject looks significantly better than it did before, and
method and fields are now handled properly.

Perhaps the most import part of this refactor is the deep integration of
__new__ and __init__ in object creation, even from the native side.