Resolved recursive method issue
The big issue up to this point was that objects had methods (obviously),
but those methods were stored as LazyValue<VMethod>. Even though these
were lazy values, the lazy value was computed by initializing an
instance of VMethod, which would infinitely recurse when trying to find
the __new__ method (since that would cause another unpacking of a lazy
value, which would search __new__ again, etc...).

The fix to this was to separate the method value types (VMethod and
VNativeMethod) with the actual method structure. To this end, there is
now a MethodWrapper, and VMethod and VNativeMethod (renamed
VBuiltinMethod and VFunction, names still pending) are just like any
other builtin type. These types do differ in their __call__ method
however; unlike every other type's implementation of __call__ (which
simply delegates to the runtime to construct the type via __new__ and
__init__), this implementation is responsible for doing the actual
calling of the method. This happens by delegating to MethodWrapper#call.