Python exceptions are no longer JVM exceptions
JVM exceptions are slow, so now instead of throwing exceptions everytime something goes wrong, we now just return an instance of VExceptionWrapper. This is an instance of VFlowStatement. This type, unfortunately, must be checked for after every call that may invoke python code (whether it is user-defined or a native method