Function nodes remember the scope they were defined in
Function nodes now have a record of the scopes that existed when they
were defined. This allows the function to access variables around it
that may no longer be in the enclosing scope of the call site.

Additionally, implemented basic global/nonlocal functionality. This
still needs further testing, as I'm not sure I fully understand the
behavior in every situation. However, at first glance it seems to behave
correctly. The only thing that definitely does not work is the
following:

    def method():
        n = n + 2

This should throw a NameError, as we are using n before it is being
defined. The only reason this error is thrown is because the engine can
tell that n is eventually going to be defined in this function, thus
making is a local. Currently, IdentifierNode#execute traverses the
entire scope chain if the variable is not global or nonlocal.

Eventually, the Scope class would need a local field for variables we
can guarantee are local, but that's a while off.